// src/interfaces/utilities.type.ts
export enum Size {
	TINY = "$6",
	SMALL = "$7",
	MEDIUM = "$8",
	BIG = "$9",
	LARGE = "$10",
}

export enum Position {
	LEFT = "left",
	RIGHT = "right",
	TOP = "top",
	BOTTOM = "bottom",
}

export enum Align {
	START = "flex-start",
	CENTER = "center",
	END = "flex-end",
}
export enum Orientation {
	VERTICAL = "vertical",
	HORIZONTAL = "horizontal",
}

export enum MeasuringUnit {
	PERCENTAGE = "%",
	CELSIUS = "°C",
	FAHRENHEIT = "F",
	VOLUME = "dB",
	LIGHT_TEMP = "K",
	HOUR = "h",
	MINUTE = "m",
	SECOND = "s",
}

export enum ClimateAction {
	HEAT = "heat",
	COOL = "cool",
}

export enum Status {
	CONNECTING = "connecting",
	IDLE = "idle",
	SUCCESS = "success",
	WARNING = "warning",
	ERROR = "error",
}